+++ src/main/java/com/github/learndifferent/mtm/mapper/TagMapper.java
@@ -52,7 +52,7 @@ public interface TagMapper {
      * @param size       size
      * @return tags
      */
-    List<String> getTagsByBookmarkId(@Param("bookmarkId") Integer bookmarkId,
+    List<String> getTagsByBookmarkId(@Param("bookmarkId") Long bookmarkId,
                                      @Param("from") int from,
                                      @Param("size") int size);
 
@@ -64,9 +64,9 @@ public interface TagMapper {
      * @param size    size
      * @return Bookmark IDs
      */
-    List<Integer> getBookmarkIdsByTagName(@Param("tagName") String tagName,
-                                          @Param("from") int from,
-                                          @Param("size") int size);
+    List<Long> getBookmarkIdsByTagName(@Param("tagName") String tagName,
+                                       @Param("from") int from,
+                                       @Param("size") int size);
 
     /**
      * Delete a tag
